Converting account links section block
