+++ blocks/src/account-links-section/editor.scss
@@ -1,0 +1,9 @@
+/**
+ * The following styles get applied inside the editor only.
+ *
+ * Replace them with your own styles or remove the file completely.
+ */
+
+.wp-block-create-block-account-invoices-section {
+	border: 1px dotted #f00;
+}
