+++ webpack.config.js
@@ -25,8 +25,14 @@ config.entry = {
     "account-profile-section",
     "index.js"
   ),
+  "account-links-section/index": path.resolve(
+    process.cwd(),
+    "blocks",
+    "src",
+    "account-links-section",
+    "index.js"
+  ),
 };
-
 config.output = {
   filename: "[name].js",
   path: path.resolve(process.cwd(), "blocks", "build"),
